backend/requirement.txt
chromadb==0.5.5
google-generativeai>=0.7.2
python-dotenv
pydantic
